管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -14,6 +14,7 @@ import constants.PropertyConst;
 import services.EmployeeService;
 
 public class EmployeeAction extends ActionBase{
+
     private EmployeeService service;
 
     @Override
@@ -26,6 +27,8 @@ public class EmployeeAction extends ActionBase{
     }
 
     public void index() throws ServletException, IOException {
+
+        if (checkAdmin()) {
 
         int page = getPage();
         List<EmployeeView> employees = service.getPerPage(page);
@@ -44,20 +47,23 @@ public class EmployeeAction extends ActionBase{
         }
 
         forward(ForwardConst.FW_EMP_INDEX);
-
+        }
     }
 
     public void entryNew() throws ServletException, IOException {
+
+        if (checkAdmin()) {
 
         putRequestScope(AttributeConst.TOKEN, getTokenId());
         putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());
 
         forward(ForwardConst.FW_EMP_NEW);
+        }
     }
 
     public void create() throws ServletException, IOException {
 
-        if (checkToken()) {
+        if (checkAdmin() && checkToken()) {
 
             EmployeeView ev = new EmployeeView(
                     null,
@@ -87,11 +93,12 @@ public class EmployeeAction extends ActionBase{
 
                 redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
             }
-
         }
     }
 
     public void show() throws ServletException, IOException {
+
+        if (checkAdmin()) {
 
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -104,9 +111,12 @@ public class EmployeeAction extends ActionBase{
         putRequestScope(AttributeConst.EMPLOYEE, ev);
 
         forward(ForwardConst.FW_EMP_SHOW);
+        }
     }
 
     public void edit() throws ServletException, IOException {
+
+        if (checkAdmin()) {
 
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -120,12 +130,13 @@ public class EmployeeAction extends ActionBase{
         putRequestScope(AttributeConst.EMPLOYEE, ev);
 
         forward(ForwardConst.FW_EMP_EDIT);
+        }
 
     }
 
     public void update() throws ServletException, IOException {
 
-        if (checkToken()) {
+        if (checkAdmin() && checkToken()) {
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
                     getRequestParam(AttributeConst.EMP_CODE),
@@ -158,13 +169,25 @@ public class EmployeeAction extends ActionBase{
 
     public void destroy() throws ServletException, IOException {
 
-        if (checkToken()) {
+        if (checkAdmin() && checkToken()) {
 
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
             putSessionScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
 
             redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+        }
+    }
+
+    private boolean checkAdmin() throws ServletException, IOException {
+
+        EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+        if (ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+        } else {
+            return true;
         }
     }
